Phase 3 Complete
Students are pre-loaded through a Plist and processed to print out
useful information to the Console.

+++ MSSubject.h
@@ -11,8 +11,8 @@
 @interface MSSubject : NSObject
 
 @property (nonatomic, strong) NSString *name;
-@property (nonatomic) int creditsEarned;
-@property (nonatomic) int creditsNeeded;
+@property (nonatomic) float creditsEarned;
+@property (nonatomic) float creditsNeeded;
 @property (strong, nonatomic) NSNumber *delayedMatriculation;
 
 @end
